n-civisense-model/requirement.txt
fastapi
uvicorn[standard]

torch
transformers
bitsandbytes
accelerate

redis
python-dotenv
rapidfuzz

openai-whisper
soundfile
ffmpeg-python

python-multipart 

pydantic
